Moves New Relic functions into helpers.request
