+++ lib/helpers.js
@@ -3,7 +3,6 @@
 const logger = require('heroku-logger');
 const Promise = require('bluebird');
 const { PhoneNumberFormat, PhoneNumberUtil } = require('google-libphonenumber');
-const newrelic = require('newrelic');
 
 const contentful = require('./contentful');
 const gambitCampaigns = require('./gambit-campaigns');
@@ -132,7 +131,7 @@ module.exports.sendResponseWithStatusCode = function (req, res, code = 200, mess
   } else {
     logger.error('sendResponseWithStatusCode', { code, response });
   }
-  exports.addNewRelicParameters(req, message);
+  helpers.request.addNewRelicParameters(req, message);
 
   return res.status(code).send(response);
 };
@@ -178,7 +177,7 @@ function sendReply(req, res, messageText, messageTemplate) {
           outbound: [req.conversation.lastOutboundMessage],
         },
       };
-      exports.addNewRelicParameters(req);
+      helpers.request.addNewRelicParameters(req);
 
       return res.send({ data });
     })
@@ -375,22 +374,4 @@ module.exports.formatMobileNumber = function formatMobileNumber(mobile, format =
   const phoneUtil = PhoneNumberUtil.getInstance();
   const phoneNumberObject = phoneUtil.parse(mobile, countryCode);
   return phoneUtil.format(phoneNumberObject, PhoneNumberFormat[format]);
-};
-
-/**
- * Posts variables from given req as New Relic custom parameters. 
- * @param {object} req
- * @param {string} responseMessage
- */
-module.exports.addNewRelicParameters = function (req, responseMessage) {
-  const paramNames = ['campaignId', 'platform', 'platformUserId', 'userId'];
-  const payload = {};
-  paramNames.forEach((paramName) => {
-    payload[paramName] = req[paramName];
-  });
-  if (responseMessage) {
-    payload.gambitResponseMessage = responseMessage;
-  }
-  logger.debug('addNewRelicParameters', payload);
-  newrelic.addCustomParameters(payload);
 };
