+++ lib/helpers/request.js
@@ -1,9 +1,15 @@
 'use strict';
 
+const logger = require('heroku-logger');
+const newrelic = require('newrelic');
+
 /**
  * Request helper
  */
 module.exports = {
+  /**
+   * Platform checks.
+   */
   isTwilio: function isTwilio(req) {
     return !!req.body.MessageSid;
   },
@@ -26,5 +32,25 @@ module.exports = {
      * @see https://www.twilio.com/docs/api/messaging/message#message-status-values
      */
     return messageStatus === 'delivered';
+  },
+  /**
+   * New Relic.
+   */
+  formatNewRelicPayload: function formatNewRelicPayload(req, gambitApiResponseMessage) {
+    const paramNames = ['campaignId', 'platform', 'platformUserId', 'userId'];
+    const payload = {};
+    paramNames.forEach((paramName) => {
+      payload[paramName] = req[paramName];
+    });
+    if (gambitApiResponseMessage) {
+      payload.gambitApiResponseMessage = gambitApiResponseMessage;
+    }
+
+    return payload;
+  },
+  addNewRelicParameters: function addNewRelicParameters(req, gambitApiResponseMessage) {
+    const payload = this.formatNewRelicPayload(req, gambitApiResponseMessage);
+    logger.debug('addNewRelicParameters', payload);
+    newrelic.addCustomParameters(payload);
   },
 };
